+++ spec/whirly_spec.rb
@@ -1,26 +1,56 @@
 require_relative "../lib/whirly"
 require "minitest/autorun"
+require "irbtools/binding"
+require "stringio"
 
 describe Whirly do
-  describe "usage" do
-    it "outputs every frame of the spinner" do
-      spinner = { "frames" => ["first", "second", "third"], "interval" => 10 }
+  before do
+    Whirly.reset
+    @capture = StringIO.new
+    Whirly.configure(non_tty: true, stream: @capture)
+  end
 
-      assert_output /first.*second.*third/m do
-        Whirly.start(spinner: spinner,  non_tty: true)
-        sleep 0.1
-        Whirly.stop
-      end
+  describe "Usage" do
+    it "outputs every frame of the spinner" do
+      spinner = { "frames" => ["first", "second", "third"], "interval" => 5 }
+
+      Whirly.start(spinner: spinner)
+      sleep 0.1
+      Whirly.stop
+
+      assert_match /first.*second.*third/m, @capture.string
     end
 
     it "calls spinner proc instead of frames if proc is given" do
-      spinner = { "proc" => ->(){ "frame" }, "interval" => 10 }
+      spinner = { "proc" => ->(){ "frame" }, "interval" => 5 }
 
-      assert_output /frame/ do
-        Whirly.start(spinner: spinner, non_tty: true)
-        sleep 0.1
-        Whirly.stop
-      end
+      Whirly.start(spinner: spinner)
+      sleep 0.1
+      Whirly.stop
+
+      assert_match /frame/, @capture.string
+    end
+  end
+
+  describe "Configure and Reset" do
+    it "can be configured before starting" do
+      Whirly.configure spinner: "dots", interval: 5
+
+      Whirly.start
+      sleep 0.1
+      Whirly.stop
+
+      assert_match /⠧/, @capture.string
+    end
+
+    it "can be reset using .reset" do
+      Whirly.configure spinner: "dots", interval: 5
+      Whirly.reset
+
+      Whirly.start(non_tty: true, stream: @capture)
+      sleep 0.1
+      Whirly.stop
+      assert_match /\A[^⠧]+\z/, @capture.string
     end
   end