+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### Next
+
+- Make paint dependency optional
+
 ### 0.1.1
 
 - `non_tty` option to force TTY behaviour (whirly deactivates itself for non TTY by default)