+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Make paint dependency optional
 - Remove pause feature
+- Separate configuring into its own method, remember whirly's configuration, can be cleared with the new .reset method
 
 ### 0.1.1